Fixed validations, welcome page design, and installed Japanese packag…

+++ app/Http/Controllers/CommentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Comment;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -34,6 +35,11 @@ class CommentController extends Controller
     public function delete($comment_id)
     {
         $delete_comment = $this->comment->findOrFail($comment_id);
+
+        //  投稿の所有者かチェック(AppServicePrividerから)
+        if (Gate::denies('isOwner', $delete_comment)) {
+            return redirect()->route('post.index');
+        }
 
         $delete_comment->delete();
